Renamed directory and package route to router
